app/requirements.txt
mysql-connector-python
psycopg2-binary
python-dotenv
